backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langgraph==0.1.4
langchain-ollama==0.1.0
apscheduler==3.10.4
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
slack_sdk==3.31.0
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
